Impart domain knowledge to Cluster initialization
Specifically, we know we'll have three clusters, because triphasic data.

More specifically, we know that the first peak occurs at the lowest end
of the data range, that the third peak occurs at the highest end of the
data range, and that the second peak is somewhere in between the
first and the third.

So, initialize the first cluster at the first element of the sorted keys
list, initialize the next cluster at the last element of the sorted keys
list, and use their integer average to initialize the final cluster.