Fix average calculation for initializeClusters
Previously, the location of the middle cluster was calculated as the
difference between the distance of the first and last clusters, then
halved.  Of course, This is not the halfway point, but rather is
somewhere closer to the origin depending on how close first and last are.

Accordingly, use the average of the first and last locations.  That is,
add them and divide by 2, rather than subtracting them and dividing by 2.