Fix decodeBitsAdvanced
Previously, there were two problems with decodeBitsAdvanced():
(1) Clusters would all update to the same location, about 3.277, and
(2) there was an endless loop somewhere.

(1) was fixed by moving the attributes inside the objects' __init__
methods, so that they don't all share the same information after having
been created.

(2) was fixed by editing the logic of the if statement converge()
so that the while loop is broken if the point assignment to the Clusters
_didn't_ change since the last assignment.