Use better test catchall
Previously, this method would return "We don't know" in cases
where the submitted morse letter was not found.  However, this
could be interpreted as part of the message.

Hence, use brackets, which are more likely to be interpreted as
not part of the original message.

Additionally, return the sequence of dots and dashes which fell through
to the last line, for debugging.