Handle short inputs better
Specifically, because the lengths of the arrays generated by splitting
differ in Java and Python, Java requires an additional call at the
end of the for loop, but only under certain circumstances.

That is, not if the length of the first ones element is 0, in which
case there are no ones, and so nextTeleFuzzy wouldn't need to be called.