Add Gaussian preference
As discussed in https://www.codewars.com/kata/decode-the-morse-code-for-real/discuss/python,
"if you have trouble discerning if the particular sequence of 1's is a dot or a dash,
assume it's a dot".  For some reason, this has been interpreted to require, at present,
magic numbers.  Hopefully others will be able to clarify the correct approach here.